Made CachingPublicKeysLoader private.
Also:
* Library now only exports FirebaseAuthTokenValidator
* FirebaseAuthTokenValidator has now private attributes except firebaseProjectId
* publicKeysLoader can't be passed to FirebaseAuthTokenValidator anymore as I'm not yet sure that this is necessary for users and don't want to risk breaking changes in the future yet
