+++ firebase_auth_token_validation/e2e_test/firebase_authentication_token_validator_test.dart
@@ -48,7 +48,7 @@ void main() {
     httpClient = CountingHttpClient();
     validator = FirebaseAuthTokenValidator(
       firebaseProjectId: projectId,
-      publicKeysLoader: CachingPublicKeysLoader.retrying(inner: httpClient),
+      inner: httpClient,
     );
   });
