+++ firebase_auth_token_validation/lib/firebase_auth_token_validation.dart
@@ -1,3 +1,3 @@
 library firebase_auth_token_validation;
 
-export 'src/firebase_auth_token_validator.dart';
+export 'src/firebase_auth_token_validator.dart' show FirebaseAuthTokenValidator;
